pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tome-mcp"
version = "0.1.0"
description = "MCP server for managing a research paper library"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "AGPL-3.0-or-later"}
keywords = [
    "mcp",
    "model-context-protocol",
    "research",
    "papers",
    "bibliography",
    "bibtex",
    "semantic-search",
    "semantic-scholar",
    "citation",
    "latex",
    "pdf",
    "literature-review",
    "ai-tools",
    "llm",
]
authors = [
    {name = "Reto Stamm"},
]
maintainers = [
    {name = "Reto Stamm"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Markup :: LaTeX",
    "Typing :: Typed",
    "Programming Language :: Python :: 3.13",
    "Environment :: Plugins",
    "Framework :: FastAPI",
]
dependencies = [
    "chromadb>=0.5.0",
    "PyMuPDF>=1.24.0",
    "bibtexparser>=2.0.0b1",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "PyYAML>=6.0",
    "rapidfuzz>=3.0.0",
    "chonkie[model2vec]>=1.0.0",
    "stop-words>=2024.1.1",
]

[project.urls]
Homepage = "https://github.com/retospect/tome-mcp"
Repository = "https://github.com/retospect/tome-mcp"
Issues = "https://github.com/retospect/tome-mcp/issues"
Changelog = "https://github.com/retospect/tome-mcp/blob/main/CHANGELOG.md"

[project.scripts]
tome-mcp = "tome.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "black>=24.0",
    "bump-my-version>=0.28.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires live external services (CrossRef, S2)",
]
filterwarnings = [
    "ignore::DeprecationWarning:onnxruntime",
    "ignore:builtin type Swig:DeprecationWarning",
    "ignore:builtin type swigvarlink:DeprecationWarning",
    "ignore:legacy embedding function config:DeprecationWarning:chromadb",
]

[tool.black]
line-length = 99
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]  # line length managed by ruff format

[tool.ruff.lint.per-file-ignores]
"tests/test_git_diff.py" = ["W293"]  # intentional trailing whitespace in diff fixtures

[tool.bumpversion]
current_version = "0.1.0"
commit = true
tag = true
tag_name = "v{new_version}"
message = "bump: {current_version} → {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = "version = \"{current_version}\""
replace = "version = \"{new_version}\""

[[tool.bumpversion.files]]
filename = "src/tome/__init__.py"
search = "__version__ = \"{current_version}\""
replace = "__version__ = \"{new_version}\""

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
